Trait noodles_core::position::SequenceIndex

pub trait SequenceIndex<T> {
    type Output: ?Sized;

    fn get(self, sequence: &[T]) -> Option<&Self::Output>;
    fn get_mut(self, sequence: &mut [T]) -> Option<&mut Self::Output>;
    fn index(self, sequence: &[T]) -> &Self::Output;
    fn index_mut(self, sequence: &mut [T]) -> &mut Self::Output;
}

A 1-based index for sequences.

Required Associated Types

type Output: ?Sized

The output returned.

Required Methods

fn get(self, sequence: &[T]) -> Option<&Self::Output>

Returns a reference to the output of the given index.

fn get_mut(self, sequence: &mut [T]) -> Option<&mut Self::Output>

Returns a mutable reference to the output of the given index.

fn index(self, sequence: &[T]) -> &Self::Output

Returns a reference to the output of the given index.

This panics if the index is out of bounds.

fn index_mut(self, sequence: &mut [T]) -> &mut Self::Output

Returns a mutable reference to the output of the given index.

This panics if the index is out of bounds.